requirements.txt
pandas
matplotlib
apache-airflow-client
gitpython
seaborn 